# music-player-backend/src/main/resources/api.yml
openapi: 3.1.1
info:
  title: Peer-Assisted Music Streaming API
  description: Backend API for hybrid prefix caching, predictive pre-fetching, and master fallback.
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /albums:
    get:
      summary: Get albums
      operationId: getAlbums
      tags:
        - Albums
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive search in album name or artist name.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: name,asc
          description: Sort as `field,direction` (e.g. `name,asc`).
      responses:
        '200':
          description: A page of albums
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumPageDto'
  /albums/{albumHash}:
    get:
      summary: Get album by hash
      operationId: getAlbumByHash
      tags:
        - Albums
      parameters:
        - name: albumHash
          in: path
          required: true
          schema:
            type: string

      responses:
        '200':
          description: Album with tracks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumExpandedDto'
        '404':
          description: Album not found
  /albums/{albumHash}/cover:
    get:
      summary: Get album cover art
      operationId: getAlbumCover
      tags:
        - Albums
      parameters:
        - name: albumHash
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cover image bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Album or cover not found
  /albums/{albumHash}/songs:
    get:
      summary: Get songs in an album
      description: Returns songs visible to the caller in the album, ordered by disc number then track number.
      operationId: getAlbumSongs
      tags:
        - Albums
      parameters:
        - name: albumHash
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of songs in the album
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
        '404':
          description: Album not found
  /artists:
    get:
      summary: Get artists
      operationId: getArtists
      tags:
        - Artists
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive search in artist name.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: name,asc
      responses:
        '200':
          description: A page of artists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistPageDto'
  /artists/{artistHash}:
    get:
      summary: Get artist by hash
      operationId: getArtistByHash
      tags:
        - Artists
      parameters:
        - name: artistHash
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artist with albums
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistExpandedDto'
        '404':
          description: Artist not found
  /artists/{artistHash}/cover:
    get:
      summary: Get artist cover art
      description: Returns the cover image of the artist's first album that has one.
      operationId: getArtistCover
      tags:
        - Artists
      parameters:
        - name: artistHash
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cover image bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Artist or cover not found
  /artists/{artistHash}/songs:
    get:
      summary: Get songs by artist
      description: Returns songs visible to the caller for this artist, ordered by song name.
      operationId: getArtistSongs
      tags:
        - Artists
      parameters:
        - name: artistHash
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of songs by the artist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
        '404':
          description: Artist not found
  /playlists:
    get:
      summary: Get the current user's playlists
      description: Returns playlists where `songFileHashes` preserves each playlist's persisted song order.
      operationId: getPlaylists
      tags:
        - Playlists
      security:
        - bearerAuth: [ ]
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive substring match on playlist name.
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: name,asc
          description: Sort as `field,direction`. Supported fields are `name` and `createdAt`.
        - name: filter[indestructible]
          in: query
          required: false
          schema:
            type: boolean
          description: When set, restricts the result to indestructible (`true`) or destructible (`false`) playlists.
        - name: includeQueue
          in: query
          required: false
          schema:
            type: boolean
          description: When true, the Queue playlist is prepended to the first page of results.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of playlists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistPageDto'
        '401':
          description: Unauthorized
    post:
      summary: Create a playlist
      description: Creates a playlist.
      operationId: createPlaylist
      tags:
        - Playlists
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaylistDto'
      responses:
        '201':
          description: Playlist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistExpandedDto'
        '401':
          description: Unauthorized
  /playlists/details-by-name/{name}:
    get:
      summary: Look up a single playlist by name
      description: |
        Returns the detail of the requesting user's playlist whose name
        matches `name` exactly. Useful for fetching the indestructible default
        playlists (e.g. `Queue`, `Favourites`) without paginating the list.
      operationId: getPlaylistByName
      tags:
        - Playlists
      security:
        - bearerAuth: [ ]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Exact case-sensitive playlist name.
      responses:
        '200':
          description: Playlist with tracks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistExpandedDto'
        '401':
          description: Unauthorized
        '404':
          description: Playlist not found
  /playlists/{playlistId}:
    get:
      summary: Get playlist by ID
      description: Returns playlist songs in persisted playlist order.
      operationId: getPlaylistById
      tags:
        - Playlists
      security:
        - bearerAuth: [ ]
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Playlist with tracks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistExpandedDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Playlist not found
    patch:
      summary: Update a playlist
      description: Updates playlist name, song order, and/or cover image. Only provided fields are updated; others remain unchanged.
      operationId: updatePlaylist
      tags:
        - Playlists
      security:
        - bearerAuth: [ ]
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlaylistDto'
      responses:
        '200':
          description: Playlist updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistExpandedDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Playlist not found
    delete:
      summary: Delete a playlist
      operationId: deletePlaylist
      tags:
        - Playlists
      security:
        - bearerAuth: [ ]
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Playlist deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Playlist not found
  /playlists/{playlistId}/cover:
    get:
      summary: Get playlist cover art
      operationId: getPlaylistCover
      tags:
        - Playlists
      security:
        - bearerAuth: [ ]
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Cover image bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Playlist or cover not found
  /playlists/{playlistId}/songs:
    get:
      summary: Get songs in a playlist
      description: Returns playlist songs visible to the caller in persisted playlist order.
      operationId: getPlaylistSongs
      tags:
        - Playlists
      security:
        - bearerAuth: [ ]
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of songs in the playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPageDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Playlist not found
...
components:
  schemas:
    ArtistDto:
      type: object
      properties:
        hash:
          type: string
        name:
          type: string
    ArtistExpandedDto:
      type: object
      description: Artist summary used in paginated list responses. Includes ordered song file hashes.
      properties:
        hash:
          type: string
        name:
          type: string
        songFileHashes:
          type: array
          items:
            type: string
          description: Ordered file hashes of all songs by this artist (ordered by song name).
    ArtistPageDto:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ArtistExpandedDto'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    AlbumDto:
      type: object
      properties:
        hash:
          type: string
        name:
          type: string
    AlbumExpandedDto:
      type: object
      description: Album summary used in paginated list responses. Includes ordered song file hashes.
      properties:
        hash:
          type: string
        name:
          type: string
        songFileHashes:
          type: array
          items:
            type: string
          description: Ordered file hashes of the album's songs (disc, track).
        durationInSeconds:
          type: integer
          description: Total duration of all songs in the album.
        artist:
          $ref: '#/components/schemas/ArtistDto'
    AlbumPageDto:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AlbumExpandedDto'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    PlaylistDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        indestructible:
          type: boolean
          default: false
          description: When true, the playlist cannot be deleted by the user.
        songFileHashes:
          type: array
          items:
            type: string
          description: Ordered song file hashes for this playlist (index equals playlist position).
    PlaylistExpandedDto:
      type: object
      description: Playlist detail used in paginated list responses. Includes song metadata.
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        indestructible:
          type: boolean
          default: false
          description: When true, the playlist cannot be deleted by the user.
        songFileHashes:
          type: array
          items:
            type: string
          description: Ordered song file hashes for this playlist (index equals playlist position).
        durationInSeconds:
          type: integer
          description: Total duration of all songs in the playlist.
    PlaylistPageDto:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PlaylistDto'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    CreatePlaylistDto:
      type: object
      required:
        - name
        - playlistSongs
      properties:
        name:
          type: string
        playlistSongs:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PlaylistSongPositionDto'
          default: [ ]
          description: Song file hashes with explicit playlist positions.
        coverImage:
          type: string
          description: Base64-encoded cover image (optional)
    UpdatePlaylistDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        playlistSongs:
          type: array
          items:
            $ref: '#/components/schemas/PlaylistSongPositionDto'
          nullable: true
          description: When present, replaces playlist songs using these explicit positions.
        coverImage:
          type: string
          nullable: true
          description: Base64-encoded cover image; empty string to remove existing cover
...
    PlaylistSongPositionDto:
      type: object
      required:
        - songFileHash
        - position
      properties:
        songFileHash:
          type: string
        position:
          type: integer
          minimum: 0
...
    SongDto:
      type: object
      properties:
        fileHash:
          type: string
        name:
          type: string
        artist:
          $ref: '#/components/schemas/ArtistDto'
        album:
          $ref: '#/components/schemas/AlbumDto'
        durationInSeconds:
          type: integer
        trackNumber:
          type: integer
        discNumber:
          type: integer
        year:
          type: integer
        lastPlayed:
          type: string
          format: date-time
          nullable: true
          description: When the requesting user last played this song; null if never.
        playCount:
          type: integer
          format: int64
          default: 0
          description: How many times the requesting user has played this song.
        likedByUser:
          type: boolean
          default: false
          description: Whether the requesting user has liked this song.
...
    SongPageDto:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SongDto'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT